项目替换为jdk17，方便后续新增ai模块

+++ meteor-chat-msg/meteor-chat-msg-api/src/main/java/com/meteor/chat/api/msg/dto/RoomMsgDTO.java
@@ -30,5 +30,5 @@ public class RoomMsgDTO implements Serializable {
     /**
      * 未读消息数量
      */
-    private Integer unreadCount;
+    private Long unreadCount;
 }
