+++ meteor-chat-msg/meteor-chat-msg-system/src/main/java/com/meteor/chat/msg/domain/vo/WSMsgMark.java
@@ -26,7 +26,7 @@ public class WSMsgMark implements Serializable {
          */
         private Integer markType;
 
-        private Integer markCount;
+        private Long markCount;
 
         private Integer actType;
     }
