+++ meteor-chat-user/meteor-chat-user-system/src/main/java/com/meteor/chat/user/adapter/WSAdapter.java
@@ -38,7 +38,7 @@ public class WSAdapter {
         return notify;
     }
 
-    public static WSBaseResp<WSFriendApply> buildFriendApply(Long uid, Integer unreadCount) {
+    public static WSBaseResp<WSFriendApply> buildFriendApply(Long uid, Long unreadCount) {
         WSFriendApply wsFriendApply = new WSFriendApply(uid, unreadCount);
         return new WSBaseResp<WSFriendApply>(WSRespTypeEnum.APPLY.getType(), wsFriendApply);
     }
