+++ meteor-chat-user/meteor-chat-user-system/src/main/java/com/meteor/chat/user/service/UserBackpackService.java
@@ -23,7 +23,7 @@ public interface UserBackpackService {
      * @param uid
      * @return
      */
-    int countRenameTimes(Long uid);
+    Long countRenameTimes(Long uid);
 
 
     /**
